+++ Engine/src/Graphics/RenderAPI.h
@@ -4,6 +4,7 @@
 #include <memory>
 
 #include "Graphics/Buffer.h"
+#include "Graphics/ConstantBuffer.h"
 #include "Graphics/Shader.h"
 
 namespace aero3d {
@@ -36,6 +37,7 @@ public:
 
     virtual std::shared_ptr<VertexBuffer> CreateVertexBuffer(BufferLayout& layout, const void* data, size_t size) = 0;
     virtual std::shared_ptr<IndexBuffer> CreateIndexBuffer(const void* data, size_t size, size_t count) = 0;
+    virtual std::shared_ptr<ConstantBuffer> CreateConstantBuffer(const void* data, size_t size) = 0;
 
     virtual std::shared_ptr<Shader> CreateShader(const char* vertexPath, const char* pixelPath) = 0;
