+++ Engine/src/Graphics/RenderCommand.cpp
@@ -52,6 +52,11 @@ std::shared_ptr<IndexBuffer> RenderCommand::CreateIndexBuffer(const void* data, 
     return m_API->CreateIndexBuffer(data, size, count);
 }
 
+std::shared_ptr<ConstantBuffer> RenderCommand::CreateConstantBuffer(const void* data, size_t size)
+{
+    return m_API->CreateConstantBuffer(data, size);
+}
+
 std::shared_ptr<Shader> RenderCommand::CreateShader(const char* vertexPath, const char* pixelPath)
 {
     return m_API->CreateShader(vertexPath, pixelPath);
