+++ Engine/src/Platform/OpenGL/OpenGLRenderAPI.cpp
@@ -4,9 +4,9 @@
 
 #include "Utils/Log.h"
 #include "Utils/Assert.h"
-#include "Platform/OpenGL/OpenGLVertexBuffer.h"
-#include "Platform/OpenGL/OpenGLIndexBuffer.h"
+#include "Platform/OpenGL/OpenGLBuffer.h"
 #include "Platform/OpenGL/OpenGLShader.h"
+#include "Platform/OpenGL/OpenGLConstantBuffer.h"
 
 namespace aero3d {
 
@@ -83,6 +83,11 @@ std::shared_ptr<VertexBuffer> OpenGLRenderAPI::CreateVertexBuffer(BufferLayout& 
 std::shared_ptr<IndexBuffer> OpenGLRenderAPI::CreateIndexBuffer(const void* data, size_t size, size_t count)
 {
     return std::make_shared<OpenGLIndexBuffer>(data, size, count);
+}
+
+std::shared_ptr<ConstantBuffer> OpenGLRenderAPI::CreateConstantBuffer(const void* data, size_t size)
+{
+    return std::make_shared<OpenGLConstantBuffer>(data, size);
 }
 
 std::shared_ptr<Shader> OpenGLRenderAPI::CreateShader(const char* vertexPath, const char* pixelPath)
