+++ Engine/src/Platform/OpenGL/OpenGLRenderAPI.h
@@ -24,6 +24,7 @@ public:
 
     virtual std::shared_ptr<VertexBuffer> CreateVertexBuffer(BufferLayout& layout, const void* data, size_t size) override;
     virtual std::shared_ptr<IndexBuffer> CreateIndexBuffer(const void* data, size_t size, size_t count) override;
+    virtual std::shared_ptr<ConstantBuffer> CreateConstantBuffer(const void* data, size_t size) override;
 
     virtual std::shared_ptr<Shader> CreateShader(const char* vertexPath, const char* pixelPath) override;
