--- Engine/src/Platform/OpenGL/OpenGLVertexBuffer.h
@@ -1,27 +1,0 @@
-#ifndef AERO3D_PLATFORM_OPENGL_OPENGLVERTEXBUFFER_H_
-#define AERO3D_PLATFORM_OPENGL_OPENGLVERTEXBUFFER_H_
-
-#include "Graphics/Buffer.h"
-
-namespace aero3d {
-
-class OpenGLVertexBuffer : public VertexBuffer
-{
-public:
-    OpenGLVertexBuffer(BufferLayout& layout, const void* data, size_t size);
-    ~OpenGLVertexBuffer();
-
-    virtual void Bind() override;
-    virtual void Unbind() override;
-
-    virtual void SetData(const void* data, size_t size) override;
-
-private:
-    unsigned int m_VBO;
-    unsigned int m_VAO;
-
-};
-
-} // namespace aero3d
-
-#endif // AERO3D_PLATFORM_OPENGL_OPENGLVERTEXBUFFER_H_
